+++ jest.config.base.js
@@ -1,6 +1,6 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
-  preset: 'ts-jest/presets/js-with-ts',
+  preset: 'ts-jest/presets/default',
   testEnvironment: 'node',
   modulePathIgnorePatterns: ['out', 'e2e_tests', 'dist'],
 };