+++ packages/language-support/src/tests/syntaxValidation/semanticValidation.test.ts
@@ -1,4 +1,4 @@
-import { setConsoleCommandsEnabled } from '../../parserWrapper';
+import { _internalFeatureFlags } from '../../featureFlags';
 import { testData } from '../testData';
 import { getDiagnosticsForQuery } from './helpers';
 
@@ -139,7 +139,7 @@ describe('Semantic validation spec', () => {
     expect(getDiagnosticsForQuery({ query })).toEqual([
       {
         message:
-          'Query cannot conclude with CALL (must be a RETURN clause, an update clause, a unit subquery call, or a procedure call with no YIELD)',
+          'Query cannot conclude with CALL (must be a RETURN clause, a FINISH clause, an update clause, a unit subquery call, or a procedure call with no YIELD).',
         offsets: {
           end: 53,
           start: 0,
@@ -374,7 +374,7 @@ describe('Semantic validation spec', () => {
     expect(getDiagnosticsForQuery({ query })).toEqual([
       {
         message:
-          'Query must conclude with a RETURN clause, an update clause, a unit subquery call, or a procedure call with no YIELD',
+          'Query must conclude with a RETURN clause, a FINISH clause, an update clause, a unit subquery call, or a procedure call with no YIELD.',
         offsets: {
           end: 25,
           start: 19,
@@ -528,7 +528,7 @@ Attempted to access graph other`,
       },
       {
         message:
-          'Query cannot conclude with MATCH (must be a RETURN clause, an update clause, a unit subquery call, or a procedure call with no YIELD)',
+          'Query cannot conclude with MATCH (must be a RETURN clause, a FINISH clause, an update clause, a unit subquery call, or a procedure call with no YIELD).',
         offsets: {
           end: 26,
           start: 17,
@@ -560,6 +560,7 @@ Attempted to access graph other`,
     expect(getDiagnosticsForQuery({ query })).toEqual([]);
   });
 
+  // This is returning errors in a different order
   test('Shows errors for COLLECT with updating subqueries', () => {
     const query = `MATCH (a)
       RETURN COLLECT { SET a.name = 1 }
@@ -764,32 +765,6 @@ Attempted to access graph other`,
     ]);
   });
 
-  test('Shows errors for variables not bound in Graph Pattern Matching', () => {
-    const query = `MATCH (a) (()--(x {prop: a.prop}))+ (b) (()--())+ (c) RETURN *`;
-
-    expect(getDiagnosticsForQuery({ query })).toEqual([
-      {
-        message: `From within a quantified path pattern, one may only reference variables, that are already bound in a previous \`MATCH\` clause.
-In this case, a is defined in the same \`MATCH\` clause as (()--(x {prop: a.prop}))+.`,
-        offsets: {
-          end: 8,
-          start: 7,
-        },
-        range: {
-          end: {
-            character: 8,
-            line: 0,
-          },
-          start: {
-            character: 7,
-            line: 0,
-          },
-        },
-        severity: 1,
-      },
-    ]);
-  });
-
   test('Accumulates errors in Graph Pattern Matching', () => {
     const query = `MATCH (p = (a)--(b))+ (p = (c)--(d))+ RETURN p`;
 
@@ -896,8 +871,7 @@ In this case, a is defined in the same \`MATCH\` clause as (()--(x {prop: a.prop
         severity: 1,
       },
       {
-        message:
-          'Type mismatch: p defined with conflicting type List<T> (expected Path)',
+        message: 'Sub-path assignment is currently not supported.',
         offsets: {
           end: 35,
           start: 23,
@@ -915,7 +889,8 @@ In this case, a is defined in the same \`MATCH\` clause as (()--(x {prop: a.prop
         severity: 1,
       },
       {
-        message: 'Sub-path assignment is currently not supported.',
+        message:
+          'Type mismatch: p defined with conflicting type List<T> (expected Path)',
         offsets: {
           end: 35,
           start: 23,
@@ -942,12 +917,12 @@ In this case, a is defined in the same \`MATCH\` clause as (()--(x {prop: a.prop
       {
         message: 'Quantified path patterns are not allowed to be nested.',
         offsets: {
-          end: 23,
+          end: 22,
           start: 16,
         },
         range: {
           end: {
-            character: 23,
+            character: 22,
             line: 0,
           },
           start: {
@@ -1004,25 +979,6 @@ In this case, a is defined in the same \`MATCH\` clause as (()--(x {prop: a.prop
       },
       {
         message:
-          'The use of shortestPath and allShortestPaths with fixed length relationships is deprecated and will be removed in a future version. Please use a path with a length of 1 [r*1..1] instead or a Match with a limit.',
-        offsets: {
-          end: 32,
-          start: 27,
-        },
-        range: {
-          end: {
-            character: 32,
-            line: 0,
-          },
-          start: {
-            character: 27,
-            line: 0,
-          },
-        },
-        severity: 2,
-      },
-      {
-        message:
           "Mixing variable-length relationships ('-[*]-') with quantified relationships ('()-->*()') or quantified path patterns ('(()-->())*') is not allowed.",
         offsets: {
           end: 32,
@@ -1039,6 +995,25 @@ In this case, a is defined in the same \`MATCH\` clause as (()--(x {prop: a.prop
           },
         },
         severity: 1,
+      },
+      {
+        message:
+          'The use of shortestPath and allShortestPaths with fixed length relationships is deprecated and will be removed in a future version. Please use a path with a length of 1 [r*1..1] instead or a Match with a limit.',
+        offsets: {
+          end: 32,
+          start: 27,
+        },
+        range: {
+          end: {
+            character: 32,
+            line: 0,
+          },
+          start: {
+            character: 27,
+            line: 0,
+          },
+        },
+        severity: 2,
       },
     ]);
   });
@@ -1096,7 +1071,7 @@ That is, neither of these is a quantified path pattern.`,
     expect(getDiagnosticsForQuery({ query })).toEqual([
       {
         message: `From within a quantified path pattern, one may only reference variables, that are already bound in a previous \`MATCH\` clause.
-In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nodes(p)[0]).prop}))*.`,
+In this case, \`p\` is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nodes(p)[0]).prop}))*.`,
         offsets: {
           end: 66,
           start: 6,
@@ -1121,15 +1096,14 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
 
     expect(getDiagnosticsForQuery({ query })).toEqual([
       {
-        message:
-          'Variable length relationships cannot be part of a quantified path pattern.',
+        message: `Mixing variable-length relationships ('-[*]-') with quantified relationships ('()-->*()') or quantified path patterns ('(()-->())*') is not allowed.`,
         offsets: {
-          end: 26,
+          end: 21,
           start: 8,
         },
         range: {
           end: {
-            character: 26,
+            character: 21,
             line: 0,
           },
           start: {
@@ -1141,14 +1115,14 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
       },
       {
         message:
-          "Mixing variable-length relationships ('-[*]-') with quantified relationships ('()-->*()') or quantified path patterns ('(()-->())*') is not allowed.",
+          'Variable length relationships cannot be part of a quantified path pattern.',
         offsets: {
-          end: 26,
+          end: 21,
           start: 8,
         },
         range: {
           end: {
-            character: 26,
+            character: 21,
             line: 0,
           },
           start: {
@@ -1200,7 +1174,7 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
     expect(getDiagnosticsForQuery({ query })).toEqual([
       {
         message:
-          'A pattern expression should only be used in order to test the existence of a pattern. It can no longer be used inside the function size(), an alternative is to replace size() with COUNT {}.',
+          'A pattern expression should only be used in order to test the existence of a pattern. It should therefore only be used in contexts that evaluate to a boolean, e.g. inside the function exists() or in a WHERE-clause. No other uses are allowed, instead they should be replaced by a pattern comprehension.',
         offsets: {
           end: 29,
           start: 22,
@@ -1233,12 +1207,12 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
       {
         message: "Label expressions are not allowed to contain '|:'.",
         offsets: {
-          end: 10,
+          end: 13,
           start: 10,
         },
         range: {
           end: {
-            character: 10,
+            character: 13,
             line: 0,
           },
           start: {
@@ -1259,12 +1233,12 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
         message:
           "Mixing the IS keyword with colon (':') between labels is not allowed. This expression could be expressed as IS A&B.",
         offsets: {
-          end: 13,
+          end: 15,
           start: 13,
         },
         range: {
           end: {
-            character: 13,
+            character: 15,
             line: 0,
           },
           start: {
@@ -1278,7 +1252,8 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
   });
 
   test('gives error on console commands when they are disabled', () => {
-    setConsoleCommandsEnabled(true);
+    _internalFeatureFlags.consoleCommands = true;
+
     expect(
       getDiagnosticsForQuery({ query: 'RETURN a;:clear; RETURN b;:history;' }),
     ).toEqual([
@@ -1319,11 +1294,11 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
         severity: 1,
       },
     ]);
-    setConsoleCommandsEnabled(false);
+    _internalFeatureFlags.consoleCommands = false;
   });
 
   test('Handles multiple cypher statements in a single query', () => {
-    setConsoleCommandsEnabled(true);
+    _internalFeatureFlags.consoleCommands = true;
     expect(getDiagnosticsForQuery({ query: 'RETURN a; RETURN b;' })).toEqual([
       {
         message: 'Variable `a` not defined',
@@ -1362,11 +1337,11 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
         severity: 1,
       },
     ]);
-    setConsoleCommandsEnabled(false);
+    _internalFeatureFlags.consoleCommands = false;
   });
 
   test('Handles cypher mixed with client commands', () => {
-    setConsoleCommandsEnabled(true);
+    _internalFeatureFlags.consoleCommands = true;
     expect(
       getDiagnosticsForQuery({
         query: ':clear;RETURN a;:clear; RETURN b;:history;',
@@ -1409,11 +1384,11 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
         severity: 1,
       },
     ]);
-    setConsoleCommandsEnabled(false);
+    _internalFeatureFlags.consoleCommands = false;
   });
 
   test('Handles cypher mixed with complex client command', () => {
-    setConsoleCommandsEnabled(true);
+    _internalFeatureFlags.consoleCommands = true;
     expect(
       getDiagnosticsForQuery({
         query: `
@@ -1444,7 +1419,7 @@ In this case, p is defined in the same \`MATCH\` clause as ((a)-[e]->(b {h: (nod
         severity: 1,
       },
     ]);
-    setConsoleCommandsEnabled(false);
+    _internalFeatureFlags.consoleCommands = false;
   });
 
   test('Does not provide semantic validation for pluggeable functions when schema is not available', () => {